chore(composables): divide useContext into useSalutations and useCountries
