+++ packages/composables/__tests__/useSalutations.spec.ts
@@ -9,7 +9,7 @@ const mockedApiClient = shopwareClient as jest.Mocked<typeof shopwareClient>;
 
 import { useContext } from "@shopware-pwa/composables";
 
-describe("Composables - useContext", () => {
+describe("Composables - useSalutations", () => {
   describe("refs", () => {
     describe("salutations", () => {
       it("should return null value when fetchSalutations is not called", async () => {
@@ -36,42 +36,6 @@ describe("Composables - useContext", () => {
         const { salutations, fetchSalutations } = useContext();
         await fetchSalutations();
         expect(salutations.value).toEqual([
-          {
-            displayName: "Mr.",
-            id: "id",
-            salutationKey: "salutatonKey",
-            letterName: "Dear Mr."
-          },
-          {
-            displayName: "Mrs.",
-            id: "id",
-            salutationKey: "salutatonKey",
-            letterName: "Dear Mrs."
-          }
-        ]);
-      });
-    });
-    describe("countries", () => {
-      it("should return salutations array", async () => {
-        mockedApiClient.getAvailableCountries.mockReturnValueOnce({
-          data: [
-            {
-              displayName: "Mr.",
-              id: "id",
-              salutationKey: "salutatonKey",
-              letterName: "Dear Mr."
-            },
-            {
-              displayName: "Mrs.",
-              id: "id",
-              salutationKey: "salutatonKey",
-              letterName: "Dear Mrs."
-            }
-          ]
-        } as any);
-        const { countries, fetchCountries } = useContext();
-        await fetchCountries();
-        expect(countries.value).toEqual([
           {
             displayName: "Mr.",
             id: "id",
@@ -116,35 +80,6 @@ describe("Composables - useContext", () => {
         ]);
       });
     });
-    describe("getMappedCoutries", () => {
-      it("should contain empty array when there aren't any available countries", () => {
-        const { getMappedCountries } = useContext();
-        expect(getMappedCountries.value).toEqual([]);
-      });
-      it("should contain properly mapped countries", () => {
-        const { countries, getMappedCountries } = useContext();
-        countries.value = [
-          {
-            name: "Norway",
-            active: true,
-            id: "id",
-            iso: "iso",
-            createdAt: "date"
-          },
-          {
-            name: "Romania",
-            active: true,
-            id: "id",
-            iso: "iso",
-            createdAt: "date"
-          }
-        ] as any;
-        expect(getMappedCountries.value).toEqual([
-          { name: "Norway", id: "id" },
-          { name: "Romania", id: "id" }
-        ]);
-      });
-    });
   });
   describe("methods", () => {
     describe("fetchSalutations", () => {
@@ -182,7 +117,6 @@ describe("Composables - useContext", () => {
             letterName: "Dear Mrs."
           }
         ]);
-        expect(shopwareClient.getAvailableCountries).toHaveBeenCalledTimes(1);
       });
       it("should assing error to error message if getAvailableSalutations throws one", async () => {
         mockedApiClient.getAvailableSalutations.mockImplementationOnce(() => {
@@ -194,57 +128,6 @@ describe("Composables - useContext", () => {
           "Error: Couldn't fetch available salutations."
         );
       });
-    });
-  });
-  describe("fetchCoutries", () => {
-    it("should fetch available countries and assign it to countries array", async () => {
-      mockedApiClient.getAvailableCountries.mockReturnValueOnce({
-        data: [
-          {
-            name: "Norway",
-            active: true,
-            id: "id",
-            iso: "iso",
-            createdAt: "date"
-          },
-          {
-            name: "Romania",
-            active: true,
-            id: "id",
-            iso: "iso",
-            createdAt: "date"
-          }
-        ]
-      } as any);
-      const { fetchCountries, countries } = useContext();
-      await fetchCountries();
-      expect(countries.value).toEqual([
-        {
-          name: "Norway",
-          active: true,
-          id: "id",
-          iso: "iso",
-          createdAt: "date"
-        },
-        {
-          name: "Romania",
-          active: true,
-          id: "id",
-          iso: "iso",
-          createdAt: "date"
-        }
-      ]);
-      expect(mockedApiClient.getAvailableCountries).toBeCalledTimes(2);
-    });
-    it("should assing error to error message if getAvailableSalutations throws one", async () => {
-      mockedApiClient.getAvailableCountries.mockImplementationOnce(() => {
-        throw new Error("Couldn't fetch available countries.");
-      });
-      const { fetchCountries, error } = useContext();
-      await fetchCountries();
-      expect(error.value.toString()).toBe(
-        "Error: Couldn't fetch available countries."
-      );
     });
   });
 });
