+++ packages/composables/src/index.ts
@@ -4,7 +4,8 @@ export * from "./hooks/useCart";
 export * from "./logic/useAddToCart";
 export * from "./hooks/useCategoryFilters";
 export * from "./hooks/useNavigation";
-export * from "./hooks/useContext";
+export * from "./hooks/useSalutations";
+export * from "./hooks/useCountries";
 export * from "./hooks/useUser";
 export * from "./hooks/useProductListing";
 export * from "./theme/cart/useCartSidebar";
